Moves Client to its own file, adds some (erroneous, see https://github.com/gravitational/teleport/pull/29659#discussion_r1276780999) comments about why its Send + Sync, renames it to FFIClient and InternalClient to RustClient. Next step will be actually making the FFIClient Send + Sync...
